ban rework & add who banned

+++ Family_and_Clans/bans.py
@@ -1,3 +1,4 @@
+import coc
 from disnake.ext import commands
 import disnake
 from utils.components import create_components
@@ -16,13 +17,8 @@ class banlists(commands.Cog, name="Bans"):
         pass
 
     @ban.sub_command(name='list', description="List of server banned players")
+    @commands.check_any(commands.has_permissions(manage_guild=True), check_commands())
     async def ban_list(self, ctx: disnake.ApplicationCommandInteraction):
-
-        perms = ctx.author.guild_permissions.manage_guild
-        if not perms:
-            embed = disnake.Embed(description="Command requires you to have `Manage Server` permissions.",
-                                  color=disnake.Color.red())
-            return await ctx.send(embed=embed)
 
         await ctx.response.defer()
 
@@ -64,10 +60,9 @@ class banlists(commands.Cog, name="Bans"):
                 for embed in embeds:
                     await ctx.channel.send(embed=embed)
 
-
     @ban.sub_command(name='add', description="Add player to server ban list")
     @commands.check_any(commands.has_permissions(manage_guild=True), check_commands())
-    async def ban_add(self, ctx: disnake.ApplicationCommandInteraction, tag: str, reason :str= "None"):
+    async def ban_add(self, ctx: disnake.ApplicationCommandInteraction, tag: str, reason: str = "None"):
         """
             Parameters
             ----------
@@ -78,63 +73,9 @@ class banlists(commands.Cog, name="Bans"):
         player = await self.bot.getPlayer(tag)
         if player is None:
             return await self.bot.player_handle(ctx, tag)
-
-
-        results = await self.bot.banlist.find_one({"$and": [
-            {"VillageTag": player.tag},
-            {"server": ctx.guild.id}
-        ]})
-
-        if results is not None and reason == "None":
-            embed = disnake.Embed(description=f"{player.name} is already banned on this server.\nProvide `reason` to update ban notes.",
-                                  color=disnake.Color.red())
-            return await ctx.send(embed=embed)
-        elif results is not None and reason != "None":
-            await self.bot.banlist.update_one({"$and": [
-                {"VillageTag": player.tag},
-                {"server": ctx.guild.id}
-            ]}, {'$set': {"Notes": reason}})
-            embed = disnake.Embed(
-                description=f"[{player.name}]({player.share_link}) ban reason updated by {ctx.author.mention}.\n"
-                            f"Notes: {reason}",
-                color=disnake.Color.green())
-            return await ctx.send(embed=embed)
-
-
-        now = datetime.now()
-        dt_string = now.strftime("%Y-%m-%d %H:%M:%S")
-
-        await self.bot.banlist.insert_one({
-            "VillageTag": player.tag,
-            "DateCreated": dt_string,
-            "Notes": reason,
-            "server": ctx.guild.id
-        })
-        embed2 = disnake.Embed(description=f"[{player.name}]({player.share_link}) added to the banlist by {ctx.author.mention}.\n"
-                                          f"Notes: {reason}",
-                               color=disnake.Color.green())
-        await ctx.send(embed=embed2)
-
-        results = await self.bot.server_db.find_one({"server": ctx.guild.id})
-        banChannel = results.get("banlist")
-        channel = await self.bot.pingToChannel(ctx,banChannel)
-
-        if channel is not None:
-            x = 0
-            async for message in channel.history(limit=200):
-                message: disnake.Message
-                x += 1
-                if x == 101:
-                    break
-                if message.author.id != self.bot.user.id:
-                    continue
-                await message.delete()
-
-            embeds = await self.create_embeds(ctx)
-            for embed in embeds:
-                await channel.send(embed=embed)
-            await channel.send(embed=embed2)
-
+        await ctx.response.defer()
+        embed = await self.ban_player(ctx, player, reason)
+        await ctx.edit_original_message(embed=embed)
 
     @ban.sub_command(name='remove', description="Remove player from server ban list")
     @commands.check_any(commands.has_permissions(manage_guild=True), check_commands())
@@ -180,7 +121,9 @@ class banlists(commands.Cog, name="Bans"):
                     break
                 if message.author.id != self.bot.user.id:
                     continue
-                await message.delete()
+                if message.embeds:
+                    if "Ban List" in str(message.embeds[0].title) or "banlist" in str(message.embeds[0].description):
+                        await message.delete()
 
             embeds = await self.create_embeds(ctx)
             for embed in embeds:
@@ -191,6 +134,68 @@ class banlists(commands.Cog, name="Bans"):
     async def clan_player_tags(self, ctx: disnake.ApplicationCommandInteraction, query: str):
         names = await self.bot.family_names(query=query, guild=ctx.guild)
         return names
+
+    async def ban_player(self, ctx, player: coc.Player, reason):
+        results = await self.bot.banlist.find_one({"$and": [
+            {"VillageTag": player.tag},
+            {"server": ctx.guild.id}
+        ]})
+
+        if results is not None and reason == "None":
+            embed = disnake.Embed(
+                description=f"{player.name} is already banned on this server.\nProvide `reason` to update ban notes.",
+                color=disnake.Color.red())
+            return embed
+        elif results is not None and reason != "None":
+            await self.bot.banlist.update_one({"$and": [
+                {"VillageTag": player.tag},
+                {"server": ctx.guild.id}
+            ]}, {'$set': {"Notes": reason, "added_by": ctx.author.id}})
+            embed = disnake.Embed(
+                description=f"[{player.name}]({player.share_link}) ban reason updated by {ctx.author.mention}.\n"
+                            f"Notes: {reason}",
+                color=disnake.Color.green())
+            return embed
+
+        now = datetime.now()
+        dt_string = now.strftime("%Y-%m-%d %H:%M:%S")
+
+        await self.bot.banlist.insert_one({
+            "VillageTag": player.tag,
+            "DateCreated": dt_string,
+            "Notes": reason,
+            "server": ctx.guild.id,
+            "added_by": ctx.author.id
+        })
+        embed2 = disnake.Embed(
+            description=f"[{player.name}]({player.share_link}) added to the banlist by {ctx.author.mention}.\n"
+                        f"Notes: {reason}",
+            color=disnake.Color.green())
+
+        results = await self.bot.server_db.find_one({"server": ctx.guild.id})
+        banChannel = results.get("banlist")
+        channel = await self.bot.pingToChannel(ctx, banChannel)
+
+        if channel is not None:
+            x = 0
+            async for message in channel.history(limit=200):
+                message: disnake.Message
+                x += 1
+                if x == 101:
+                    break
+                if message.author.id != self.bot.user.id:
+                    continue
+                if message.embeds:
+                    if "Ban List" in str(message.embeds[0].title) or "banlist" in str(message.embeds[0].description):
+                        await message.delete()
+
+            embeds = await self.create_embeds(ctx)
+            for embed in embeds:
+                await channel.send(embed=embed)
+            await channel.send(embed=embed2)
+
+        return embed2
+
 
     async def create_embeds(self, ctx):
         text = []
@@ -219,10 +224,14 @@ class banlists(commands.Cog, name="Bans"):
                 clan = f"{clan}, {str(player.role)}"
             except:
                 clan = "No Clan"
+            added_by = ""
+            if ban.get("added_by") is not None:
+                user = await self.bot.getch_user(ban.get("added_by"))
+                added_by = f"\nAdded by: {user}"
             hold += f"{emojiDictionary(player.town_hall)}[{name}]({player.share_link}) | {player.tag}\n" \
                     f"{clan}\n" \
                     f"Added on: {date}\n" \
-                    f"*{notes}*\n\n"
+                    f"Notes: *{notes}*{added_by}\n\n"
             num += 1
             if num == 10:
                 text.append(hold)
